// front/lib/generated/asset_registry/types/vaultVoteOutcome.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum VaultVoteOutcome {
  Pending,
  ReturnCapital,
  Extend,
}

export type VaultVoteOutcomeArgs = VaultVoteOutcome;

export function getVaultVoteOutcomeEncoder(): FixedSizeEncoder<VaultVoteOutcomeArgs> {
  return getEnumEncoder(VaultVoteOutcome);
}

export function getVaultVoteOutcomeDecoder(): FixedSizeDecoder<VaultVoteOutcome> {
  return getEnumDecoder(VaultVoteOutcome);
}

export function getVaultVoteOutcomeCodec(): FixedSizeCodec<
  VaultVoteOutcomeArgs,
  VaultVoteOutcome
> {
  return combineCodec(
    getVaultVoteOutcomeEncoder(),
    getVaultVoteOutcomeDecoder(),
  );
}