// front/lib/generated/asset_registry/types/payoutVaultState.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum PayoutVaultState {
  Active,
  Frozen,
  Completed,
  Cancelled,
}

export type PayoutVaultStateArgs = PayoutVaultState;

export function getPayoutVaultStateEncoder(): FixedSizeEncoder<PayoutVaultStateArgs> {
  return getEnumEncoder(PayoutVaultState);
}

export function getPayoutVaultStateDecoder(): FixedSizeDecoder<PayoutVaultState> {
  return getEnumDecoder(PayoutVaultState);
}

export function getPayoutVaultStateCodec(): FixedSizeCodec<
  PayoutVaultStateArgs,
  PayoutVaultState
> {
  return combineCodec(
    getPayoutVaultStateEncoder(),
    getPayoutVaultStateDecoder(),
  );
}